+++ src/NHSD.GPIT.BuyingCatalogue.ServiceContracts/Solutions/IManageFiltersService.cs
@@ -19,8 +19,6 @@ namespace NHSD.GPIT.BuyingCatalogue.ServiceContracts.Solutions
 
         Task<bool> FilterExists(string filterName, int organisationId);
 
-        Task<Filter> GetFilter(int organisationId, int filterId);
-
         Task<List<Filter>> GetFilters(int organisationId);
 
         Task<FilterDetailsModel> GetFilterDetails(int organisationId, int filterId);
