+++ src/NHSD.GPIT.BuyingCatalogue.WebApp/Areas/Competitions/Views/CompetitionsDashboard/ReviewFilter.cshtml
@@ -1,9 +1,6 @@
 @using Microsoft.AspNetCore.Mvc.TagHelpers
-@using NHSD.GPIT.BuyingCatalogue.EntityFramework.Extensions
 @using NHSD.GPIT.BuyingCatalogue.UI.Components.TagHelpers
-@using NHSD.GPIT.BuyingCatalogue.UI.Components.Views.Shared.TagHelpers.SummaryList
-@using NHSD.GPIT.BuyingCatalogue.UI.Components.Views.Shared.TagHelpers.Table
-@model ReviewFilterModel
+@model NHSD.GPIT.BuyingCatalogue.WebApp.Models.Shared.ReviewFilterModel
 
 @{
     ViewBag.Title = "Review your filter";
@@ -17,140 +14,7 @@
                         caption=@Model.FilterDetails.Name
                         advice="These are the details for this filter."/>
 
-        <h2 data-test-id="name-and-description">Name and description</h2>
-
-        <nhs-summary-list>
-            <nhs-summary-list-row label-text="Name">
-                <p>
-                    @Model.FilterDetails.Name
-                </p>
-            </nhs-summary-list-row>
-            <nhs-summary-list-row label-text="Description">
-                <p>
-                    @Model.FilterDetails.Description
-                </p>
-            </nhs-summary-list-row>
-        </nhs-summary-list>
-
-        @if (Model.HasEpics())
-        {
-            <h2 data-test-id="capabilities-and-epics">Capabilities and Epics</h2>
-            <nhs-table>
-                <nhs-table-column>Capability</nhs-table-column>
-                <nhs-table-column>Epic</nhs-table-column>
-                @foreach (var capability in Model.FilterDetails.Capabilities)
-                {
-                    <nhs-table-row-container>
-                        <nhs-table-cell>
-                            @capability.Key
-                        </nhs-table-cell>
-                        <nhs-table-cell>
-                            @if (capability.Value is null || !capability.Value.Any())
-                            {
-                                <p>
-                                    None selected
-                                </p>
-                            }
-                            else
-                            {
-                                @if (capability.Value.Count() == 1)
-                                {
-                                    <p>
-                                        @capability.Value.First()
-                                    </p>
-                                }
-                                else
-                                {
-                                    @foreach (var epic in capability.Value)
-                                    {
-                                        <li>
-                                            @epic
-                                        </li>
-                                    }
-                                }
-                            }
-                        </nhs-table-cell>
-                    </nhs-table-row-container>
-                }
-            </nhs-table>
-        }
-        else
-        {
-            <h2 data-test-id="capabilities">Capabilities</h2>
-            @foreach (var capability in Model.FilterDetails.Capabilities)
-            {
-                <p>@capability.Key</p>
-            }
-        }
-
-        @if (Model.HasAdditionalFilters())
-        {
-            <h2 data-test-id="additional-filters">Additional filters</h2>
-
-            <nhs-summary-list>
-                <nhs-summary-list-row label-text="Client application type">
-                    @if (!Model.HasClientApplicationTypes())
-                    {
-                        <p>
-                            None selected
-                        </p>
-                    }
-                    else
-                    {
-                        @if (Model.FilterDetails.ClientApplicationTypes.Count == 1)
-                        {
-                            <p>
-                                @Model.FilterDetails.ClientApplicationTypes.First().Name()
-                            </p>
-                        }
-                        else
-                        {
-                            <ul>
-                                @foreach (var item in Model.FilterDetails.ClientApplicationTypes)
-                                {
-                                    <li>
-                                        @item.Name()
-                                    </li>
-                                }
-                            </ul>
-                        }
-                    }
-                </nhs-summary-list-row>
-                <nhs-summary-list-row label-text="Framework">
-                    <p>
-                        @(Model.HasFramework() ? Model.FilterDetails.FrameworkName : "None selected")
-                    </p>
-                </nhs-summary-list-row>
-                <nhs-summary-list-row label-text="Hosting type">
-                    @if (!Model.HasHostingTypes())
-                    {
-                        <p>
-                            None selected
-                        </p>
-                    }
-                    else
-                    {
-                        @if (Model.FilterDetails.HostingTypes.Count == 1)
-                        {
-                            <p>
-                                @Model.FilterDetails.HostingTypes.First().Name()
-                            </p>
-                        }
-                        else
-                        {
-                            <ul>
-                                @foreach (var item in Model.FilterDetails.HostingTypes)
-                                {
-                                    <li>
-                                        @item.Name()
-                                    </li>
-                                }
-                            </ul>
-                        }
-                    }
-                </nhs-summary-list-row>
-            </nhs-summary-list>
-        }
+        <partial name="Partials/_FilterDetailsPartial" model="Model" />
 
         <form method="post">
             <nhs-submit-button text="Continue"/>
