+++ tests/NHSD.GPIT.BuyingCatalogue.Services.UnitTests/Solutions/ManageFiltersServiceTests.cs
@@ -586,39 +586,5 @@ namespace NHSD.GPIT.BuyingCatalogue.Services.UnitTests.Solutions
             filterDetails.Description.Should().Be(filter.Description);
             filterDetails.FrameworkName.Should().Be(framework.ShortName);
         }
-
-        [Theory]
-        [InMemoryDbAutoData]
-        public static async Task GetFilter_ReturnsExpectedResults(
-            Filter filter,
-            Organisation organisation,
-            EntityFramework.Catalogue.Models.Framework framework,
-            [Frozen] BuyingCatalogueDbContext context,
-            ManageFiltersService service)
-        {
-            context.Organisations.Add(organisation);
-            await context.SaveChangesAsync();
-
-            context.Frameworks.Add(framework);
-            await context.SaveChangesAsync();
-
-            filter.Organisation = organisation;
-            filter.OrganisationId = organisation.Id;
-            filter.Framework = framework;
-            filter.FrameworkId = framework.Id;
-
-            await context.SaveChangesAsync();
-
-            context.Filters.Add(filter);
-            await context.SaveChangesAsync();
-            context.ChangeTracker.Clear();
-
-            var result = await service.GetFilter(organisation.Id, filter.Id);
-
-            result.Should().NotBeNull();
-            result.Id.Should().Be(filter.Id);
-            result.OrganisationId.Should().Be(filter.OrganisationId);
-            result.Framework.Should().BeEquivalentTo(filter.Framework);
-        }
     }
 }
